+++ tests/test_transform.py
@@ -54,22 +54,16 @@ def test_as_freq_daily(il_electricity_cdd_hdd_billing_monthly):
     meter_data = il_electricity_cdd_hdd_billing_monthly["meter_data"]
     assert meter_data.shape == (27, 1)
     as_daily = as_freq(meter_data.value, freq="D")
-    assert as_daily.shape == (791,)
-    assert (
-        round(meter_data.value.sum(), 1) == round(as_daily.sum(), 1) + 12.0 == 21290.2
-    )
+    assert as_daily.shape == (792,)
+    assert round(meter_data.value.sum(), 1) == round(as_daily.sum(), 1) == 21290.2
 
 
 def test_as_freq_month_start(il_electricity_cdd_hdd_billing_monthly):
     meter_data = il_electricity_cdd_hdd_billing_monthly["meter_data"]
     assert meter_data.shape == (27, 1)
     as_month_start = as_freq(meter_data.value, freq="MS")
-    assert as_month_start.shape == (27,)
-    assert (
-        round(meter_data.value.sum(), 1)
-        == round(as_month_start.sum(), 1) + 925.0
-        == 21290.2
-    )
+    assert as_month_start.shape == (28,)
+    assert round(meter_data.value.sum(), 1) == round(as_month_start.sum(), 1) == 21290.2
 
 
 def test_as_freq_hourly_temperature(il_electricity_cdd_hdd_billing_monthly):
@@ -84,7 +78,7 @@ def test_as_freq_daily_temperature(il_electricity_cdd_hdd_billing_monthly):
     temperature_data = il_electricity_cdd_hdd_billing_monthly["temperature_data"]
     assert temperature_data.shape == (19417,)
     as_daily = as_freq(temperature_data, freq="D", series_type="instantaneous")
-    assert as_daily.shape == (810,)
+    assert as_daily.shape == (811,)
     assert abs(temperature_data.mean() - as_daily.mean()) <= 0.1
 
 
@@ -92,8 +86,8 @@ def test_as_freq_month_start_temperature(il_electricity_cdd_hdd_billing_monthly)
     temperature_data = il_electricity_cdd_hdd_billing_monthly["temperature_data"]
     assert temperature_data.shape == (19417,)
     as_month_start = as_freq(temperature_data, freq="MS", series_type="instantaneous")
-    assert as_month_start.shape == (28,)
-    assert round(as_month_start.mean(), 1) == 54.5
+    assert as_month_start.shape == (29,)
+    assert round(as_month_start.mean(), 1) == 53.4
 
 
 def test_as_freq_daily_temperature_monthly(il_electricity_cdd_hdd_billing_monthly):
@@ -109,6 +103,21 @@ def test_as_freq_empty():
     meter_data = pd.DataFrame({"value": []})
     empty_meter_data = as_freq(meter_data.value, freq="H")
     assert empty_meter_data.empty
+
+
+def test_as_freq_perserves_nulls(il_electricity_cdd_hdd_billing_monthly):
+    meter_data = il_electricity_cdd_hdd_billing_monthly["meter_data"]
+    monthly_with_nulls = meter_data[meter_data.index.year != 2016].reindex(
+        meter_data.index
+    )
+    daily_with_nulls = as_freq(monthly_with_nulls.value, freq="D")
+    assert (
+        round(monthly_with_nulls.value.sum(), 2)
+        == round(daily_with_nulls.sum(), 2)
+        == 11094.05
+    )
+    assert monthly_with_nulls.value.isnull().sum() == 13
+    assert daily_with_nulls.isnull().sum() == 365
 
 
 def test_day_counts(il_electricity_cdd_hdd_billing_monthly):